feat(select): improving types and splitting into seperate files
