+++ src/components/FormElements/Select/Select.tsx
@@ -1,142 +1,23 @@
 // Inspired by this https://codesandbox.io/embed/m75wlyx3oy
 
-import React, {
-  Children,
-  ForwardRefRenderFunction,
-  forwardRef,
-  useEffect,
-  useRef,
-  useState,
-  MouseEvent as ReactMouseEvent,
-} from "react";
-import ReactSelect, { CommonProps, GroupBase, SelectInstance, components } from "react-select";
+import React, { ForwardRefRenderFunction, forwardRef, useEffect, useRef, useState } from "react";
+import ReactSelect, { SelectInstance } from "react-select";
 import { SerializedStyles } from "@emotion/react";
 import classNames from "classnames";
-import { searchInputContainer } from "../Input/styles";
 import Label from "../Label/Label";
-import Input from "../Input/Input";
-import { customMenuList, selectContainer } from "./styles";
-import {
-  CustomMenuListProps,
-  CustomOptionType,
-  CustomSelectProps,
-  CustomValueContainerProps,
-} from "./types";
+import CustomValueContainer from "./components/CustomValueContainer";
+import { selectContainer } from "./styles";
+import CustomMenuList from "./components/CustomMenuList";
+import { CustomOption, CustomSelectProps } from "./types";
 import { formElements } from "@theme/default/config";
 
 const MAX_MENU_HEIGHT = 300;
 const OUTER_PLACEHOLDER = "Select...";
 const INNER_PLACEHOLDER = "Search...";
 
-const { MenuList, ValueContainer, SingleValue, Placeholder } = components;
-
-const CustomMenuList: React.FC<CustomMenuListProps<CustomOptionType>> = (customMenuProps) => {
-  const { selectProps, ...props } = customMenuProps;
-  const { onInputChange, inputValue, innerPlaceholder } = selectProps;
-  const { onMenuInputFocus } = selectProps;
-
-  const ariaAttributes = {
-    "aria-autocomplete": "list" as const,
-    "aria-label": selectProps["aria-label"],
-    "aria-labelledby": selectProps["aria-labelledby"],
-  };
-
-  return (
-    <div css={customMenuList}>
-      <div css={searchInputContainer} onMouseDown={(e) => e.stopPropagation()}>
-        <Input
-          id="react-select-inner-search-input"
-          placeholder={innerPlaceholder}
-          autoCorrect="off"
-          autoComplete="off"
-          spellCheck="false"
-          type="text"
-          value={inputValue}
-          showVerticalLine={false}
-          isClearable={true}
-          onChange={(e) => {
-            onInputChange(e.target.value, {
-              action: "input-change",
-              prevInputValue: inputValue,
-            });
-          }}
-          onMouseDown={(e: ReactMouseEvent<HTMLInputElement>) => {
-            e.stopPropagation();
-            const input = e.target as HTMLInputElement;
-            input.focus();
-          }}
-          onClear={() => {
-            onInputChange("", {
-              action: "input-change",
-              prevInputValue: inputValue,
-            });
-          }}
-          onFocus={onMenuInputFocus}
-          {...ariaAttributes}
-        />
-      </div>
-
-      <MenuList {...props} selectProps={selectProps} />
-    </div>
-  );
-};
-
-const CustomValueContainer: React.FC<CustomValueContainerProps<CustomOptionType>> = (
-  customProps,
-) => {
-  const { children, selectProps, ...props } = customProps;
-  const { isFocused = false } = selectProps;
-
-  const commonProps: CommonProps<CustomOptionType, false, GroupBase<CustomOptionType>> = {
-    clearValue: props.clearValue,
-    getStyles: props.getStyles,
-    getValue: props.getValue,
-    hasValue: props.hasValue,
-    isMulti: props.isMulti,
-    isRtl: props.isRtl,
-    options: props.options,
-    selectOption: props.selectOption,
-    setValue: props.setValue,
-    selectProps,
-    cx: props.cx,
-    getClassNames: props.getClassNames,
-    theme: props.theme,
-  };
-
-  return (
-    <ValueContainer {...props} selectProps={selectProps}>
-      {Children.map(children, (child) => {
-        return child ? (
-          child
-        ) : props.hasValue ? (
-          <SingleValue
-            {...commonProps}
-            isDisabled={selectProps.isDisabled}
-            getClassNames={props.getClassNames}
-            innerProps={Object.assign({}, props.innerProps, { "data-role": "menuList" })}
-            data={props.getValue()[0]}
-          >
-            {selectProps.getOptionLabel(props.getValue()[0])}
-          </SingleValue>
-        ) : (
-          <Placeholder
-            {...commonProps}
-            isDisabled={selectProps.isDisabled}
-            getClassNames={props.getClassNames}
-            innerProps={Object.assign({}, props.innerProps, { "data-role": "menuList" })}
-            isFocused={isFocused}
-          >
-            {selectProps.placeholder}
-          </Placeholder>
-        );
-      })}
-    </ValueContainer>
-  );
-};
-
-const CustomSelect: ForwardRefRenderFunction<
-  SelectInstance<CustomOptionType>,
-  CustomSelectProps<CustomOptionType>
+const Select: ForwardRefRenderFunction<
+  SelectInstance<CustomOption>,
+  CustomSelectProps<CustomOption>
 > = (props, forwardedRef) => {
   const {
     id = "",
@@ -147,8 +28,8 @@ const CustomSelect: ForwardRefRenderFunction<
     status = "valid",
     maxMenuHeight = MAX_MENU_HEIGHT,
     hasInnerSearch = false,
-    placeholder: outerPlaceholder = OUTER_PLACEHOLDER,
     innerPlaceholder = INNER_PLACEHOLDER,
+    placeholder: outerPlaceholder = OUTER_PLACEHOLDER,
     ...rest
   } = props;
   const hasLabel = Boolean(label);
@@ -220,26 +101,26 @@ const CustomSelect: ForwardRefRenderFunction<
           }}
           components={{
             IndicatorSeparator: () => null,
-            MenuList: hasInnerSearch ? CustomMenuList : MenuList,
-            ValueContainer: (props) => CustomValueContainer(props),
+            MenuList: CustomMenuList,
+            ValueContainer: (props) => CustomValueContainer({ ...props, isFocused }),
           }}
           {...{
-            menuIsOpen: isFocused || undefined,
-            isFocused: isFocused,
             onMenuInputFocus: () => setIsFocused(true),
             onMouseDown: (e: MouseEvent) => e.stopPropagation(),
-            innerPlaceholder,
           }}
+          menuIsOpen={isFocused || undefined}
           inputValue={inputValue}
           onChange={() => setIsFocused(false)}
           onInputChange={(val) => {
             setInputValue(val);
           }}
           blurInputOnSelect={true}
+          innerPlaceholder={innerPlaceholder}
+          hasInnerSearch={hasInnerSearch}
         />
       </div>
     </div>
   );
 };
 
-export default forwardRef(CustomSelect);
+export default forwardRef(Select);
